ensure that we start logging from the beginning of the line
This is related to https://github.com/ember-cli/ember-cli-deploy/pull/280

In order to ensure proper rendering we need to reset the line position
to 0. This can not be the case if the progress bar is rendering and a
plugin tries to display a log message (i.e. the redis plugin notifying
about the upoaded revision)